Bug Fix 🐛 	modified:   Cargo.lock 	modified:   Cargo.toml 	modified:   README.md 	modified:   src/lib.rs 	modified:   windows-capture-python/Cargo.toml 	modified:   windows-capture-python/README.md 	modified:   windows-capture-python/pyproject.toml 	modified:   windows-capture-python/windows_capture/__init__.py 	modified:   windows-capture-python/windows_capture/windows_capture.cp311-win_amd64.pyd

+++ windows-capture-python/windows_capture/__init__.py
@@ -229,7 +229,7 @@ class WindowsCapture:
         if self.frame_handler:
             internal_capture_control = InternalCaptureControl(stop_list)
 
-            row_pitch = buf_len / height
+            row_pitch = int(buf_len / height)
             if row_pitch == width * 4:
                 ndarray = numpy.ctypeslib.as_array(
                     ctypes.cast(buf, ctypes.POINTER(ctypes.c_uint8)),
